experimentos/nano-banana-test/requirements.txt
# Dependencias para Nano Banana Test
# Instalación: pip install -r requirements.txt

# Google GenAI SDK (nuevo SDK unificado para Gemini + Imagen)
google-genai>=0.3.0

# Procesamiento de imágenes
Pillow>=10.0.0

# Gestión de variables de entorno (opcional, el script funciona sin esto)
python-dotenv>=1.0.0

# ============================================================
# 📦 INSTALACIÓN RECOMENDADA:
# ============================================================
# python -m venv venv
# source venv/bin/activate  # En Windows: venv\Scripts\activate
# pip install -r requirements.txt
